+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 - PR #287 Enabled CUDA CXX11 for unit tests.
 - PR #292 Revamped RMM exceptions.
 - PR #162 New suballocator memory_resources.
+- PR #314 Improve pool_memory_resource growth strategy.
+- PR #319 Add thread_safe_resource_adaptor class.
 
 ## Improvements